Simpler configuration
The original configuration uses stack.py and supports individual minion
customizations.  With a default and user pillar directory tree, this led
to confusion for many more than once.

This strategy abandons using stack.py and the custom trees.  Instead of
file globbing, values of policy.cfg become Salt targets.  The roles are
generated the same without the intermittent files.  Lastly, the
global.yml moves out of the stack directory and is only initialized.
The global.yml is still a user maintained file.

Signed-off-by: Eric Jackson <ejackson@suse.com>